+++ apps/mobile/package.json
@@ -62,7 +62,6 @@
     "react-native-svg": "^15.11.2",
     "react-native-webview": "^13.13.5",
     "tailwind-merge": "^2.2.1",
-    "use-debounce": "^10.0.0",
     "zod": "^3.24.2",
     "zustand": "^5.0.5"
   },
